requirements.txt
azure-functions
pytz
textblob
requests
